arc_res_edit/requirements.txt
pydub
moviepy
Pillow
numpy